+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Prevent session prototype methods from being overwritten
+
 1.8.0 / 2014-09-07
 ==================